requirements.txt
dalle2-pytorch==1.15.6
filelock==3.15.4
fsspec==2024.6.1
jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
open-clip-torch==2.26.1
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pytorch-warmup==0.1.1
resize-right==0.0.2
rotary-embedding-torch==0.6.4
sympy==1.13.0
torch==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
typing-extensions==4.12.2
vector-quantize-pytorch==1.15.2
webdataset==0.2.86
x-clip==0.14.4
